Refactor fst features
Having features stored in the fst module is confusing as user don't know
where to find them. Even more now that the FstNeuron class does not exist
anymore.

So:
- features have now their own module
- Added colorized output to './morph_stats -l' which list the list
of features.
- './morph_stats -l' now accept an optional argument to filter the list of
features
