+++ apps/morph_stats
@@ -70,8 +70,8 @@ def get_parser():
 
     parser.add_argument('-v', '--verbose', action='count', dest='verbose', default=0,
                         help='-v for INFO, -vv for DEBUG')
-    parser.add_argument('-l', '--list', action='store_true', default=False,
-                        help='List the available features')
+    parser.add_argument('-l', '--list', nargs='?', const='',
+                        help='List the available features. An optional pattern can be passed to filter the results')
 
     parser.add_argument('--as-population',
                         action='store_true',
@@ -133,8 +133,9 @@ if __name__ == '__main__':
                                logging.INFO,
                                logging.DEBUG)[min(_args.verbose, 2)])
 
-    if _args.list:
-        print(nm.fst._get_doc())  # pylint: disable=W0212
+
+    if _args.list is not None:
+        print(nm.features._get_doc(_args.list))  # pylint: disable=W0212
         sys.exit(0)
     elif not _args.datapath:
         get_parser().print_usage()
