+++ neurom/features/_bifurcationfunc.py
